Creation od LXMLHandler.py class
